docs(asse-d): #1928 DEC-B-8 explicit column TestRunId vs shadow
Sessione 40 fase 1 — T1 Integration-trait re-implementation in
apps/api/tests/Api.Tests/Integration/Testing/ con IntegrationWebApplicationFactory
+ EnsureCreatedAsync. Scoperti + risolti 3 sub-blocker progressivi:
1. Shadow property su User aggregate trigger BackupCode ctor regression →
   FIX registrazione su UserEntity persistence model
2. Shadow property handler usa xxxEntity (persistence) ma OnModelCreating
   usa xxx domain → FIX tutte le 5 shadow su persistence entities
3. MigrateAsync non include shadow in schema → FIX EnsureCreatedAsync test

4° blocker NON risolto fase 1: shadow property value null dopo SaveChanges +
re-fetch (EF Core 9 + Npgsql interaction unclear).

DEC-B-8 LOCKATA: replace shadow property con explicit column TestRunId su
5 persistence entities (production-proven pattern, no EF Core gotchas).
Sessione 40 fase 2 implementa DEC-B-8 (~30-45min effort).

Files staged sessione 40 fase 1:
- MeepleAiDbContext.cs (5 shadow property su persistence entities)
- Integration/Testing/SeedTestGameNightCommandHandlerTests.cs (new file)
- Unit/Testing/SeedTestGameNightCommandHandlerTests.cs (deleted, replaced by Integration)
- spec doc DEC-B-8 addendum

Refs #1928
